Rename errorRetry.dart to error_retry.dart for file_names lint
